+++ branches/slurm-1.1/src/plugins/select/bluegene/block_allocator/block_allocator.h
@@ -264,7 +264,7 @@ extern void ba_fini();
  * IN ba_node: ba_node_t to put down
  * IN state: new state of ba_node_t
  */
-extern void ba_update_node_state(ba_node_t *ba_node, int state);
+extern void ba_update_node_state(ba_node_t *ba_node, uint16_t state);
 
 /** 
  * copy info from a ba_node
